Change style of navbar to be relatively positioned
This is part of the work to make the navbar relatively positioned
instead of absolutely positioned.

Navigation menu and search menu are now above relatively positioned
instead of relatively positioned. The rules for positioning the menus
have been removed as they are now no longer needed. Instead we will use
the govuk-frontend class `govuk-width-container` in the partial to use
the correct margins and spacing. This means there is no longer any
inconsistency between the margins of the menus and the margins of the page.

The buttons and links are now absolutely positioned instead of
relatively positioned. To ensure correct positioning, they will reside
within a `govuk-width-container` in the partial.

The styling for the search menu and navbar menu containers have been
consolidated into one class. Instead the search menu content container
and navbar menu container will apply different padding rules. This is to
ensure the padding for both menus is consistent.